refactor(backend): standardize Comment constructs names to Model, and Type
